feat(domain): encapsulate price in a Money value object
Prices were plain numbers, which invites float precision errors. Product.price is
now a Money value object backed by integer cents, so every in-domain comparison is
exact. Conversions to and from a decimal happen only at the edges (Postgres
Decimal, Elasticsearch scaled_float, HTTP JSON), so ProductProps and every
projection keep their plain-number contract and the change stays contained.

- New Money VO (fromDecimal/fromCents, toDecimal/toCents, comparisons) with tests
  covering rounding and the classic float pitfalls.
- Product builds price via Money.fromDecimal; toPrimitives/mappers convert back.

+++ src/products/domain/product.spec.ts
@@ -19,7 +19,7 @@ describe('Product.create', () => {
   it('builds a product from valid props', () => {
     const product = Product.create(props());
     expect(product.name).toBe('Laptop');
-    expect(product.price).toBe(999.99);
+    expect(product.price.toDecimal()).toBe(999.99);
   });
 
   it('rejects a blank name', () => {
